update system json to make better defaults
Whichever subtype is first listed under each documentType is selected by default when making a new item.
Rule is selected for item creation as they are commonly made in higher number than playbooks,
And pilot is selected for actor creation for the time being. Once we get NPC types they
should likely be the default for actor creation